git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@2159 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWorkSim/src/rwsim/simulator/rwphysics/RWBodyPool.cpp
@@ -36,7 +36,7 @@ RWBody *RWBodyPool::createBody(RWBody::BodyType type){
 void RWBodyPool::deleteBody(RWBody* body){
 	// todo: check if body is allready deleted
 	int bidx = body->getId();
-	RW_ASSERT(0<=bidx && bidx<_bodies.size());
+	RW_ASSERT(0<=bidx && bidx<(int)_bodies.size());
 	_freeBodyIDs.push(bidx);
 }
